2/07/24
Reading Datafile:
The script reads the CSV file using Pandas read_csv() function. Different encoding schemes (utf-8, utf-8-sig, latin1, ISO-8859-1) are tried until the file is successfully read without encountering a UnicodeDecodeError.
The first 20 rows of the DataFrame are displayed to inspect the data.
Sentiment Analysis:
Opinion Lexicon:

NLTK's opinion lexicon dataset is downloaded, which contains lists of positive and negative opinion words.
The total number of words in the opinion lexicon, as well as examples of positive and negative words, are printed.
Bing Liu Score Calculation:

A function named bing_liu_score() is defined to calculate sentiment scores for the text.
Positive words are assigned a score of 1, and negative words are assigned a score of -1.
The function tokenizes the input text into words, converts them to lowercase, and checks each word against the sentiment dictionary. If a word is found in the dictionary, its sentiment score is added to the overall sentiment score.
The function returns the sentiment score for the input text.
Applying Sentiment Analysis:

The sentiment analysis function is applied to the 'athlete' column (renamed from 'reviewText') of the DataFrame to calculate sentiment scores for each athlete's text.
NaN values in the 'athlete' column are filled with 'no review'.
The calculated sentiment scores are stored in a new column named 'Bing_Liu_Score' in the DataFrame.